Start all strings off with fresh literal types.
All string literals will now have their types inferred as string literal types with an associated "freshness".
When a string is bound to a "read-only" declaration, freshness associated with the type is lost during widening.
When a string is bound to a mutable declaration, it is unconditionally widened to 'string.
In any other context where widening takes place, any fresh string literal type is widened to 'string.